Move getInputChanges() to IncrementalTaskAction
