+++ subprojects/core/src/main/java/org/gradle/api/internal/project/taskfactory/package-info.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2017 the original author or authors.
+ * Copyright 2018 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,13 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-
+@NonNullApi
 package org.gradle.api.internal.project.taskfactory;
 
-import org.gradle.api.Action;
-import org.gradle.api.Task;
-import org.gradle.internal.reflect.Instantiator;
-
-public interface TaskActionFactory  {
-    Action<? super Task> create(Instantiator instantiator);
-}
+import org.gradle.api.NonNullApi;
